Adding Product state globally
